requirements.txt
# SQL Generation Pipeline Dependencies

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.66.4
python-dotenv>=1.0.0

# API and HTTP
httpx>=0.27.0
litellm>=1.0.0
openai>=1.34.0

# Database connections
psycopg2-binary>=2.9.9
pymysql>=1.1.1
sqlite3>=0.0.0

# Vector database and embeddings
lancedb>=0.3.0
sentence-transformers>=2.2.2

# Graph processing
networkx>=2.8.0

# Utilities
pydantic>=2.7.4
func-timeout>=4.3.5
